Fix bug mother lookup event creating woman registration events
This fixes an issue where when you are registering a child with
an existing mother a new Mother Registration event is created.

A mother lookup attribute is appended to the json form when lookup is performed. We use
this property to determine whether this child is a sibling to another child. If so we return
null for the mother registraiton event.

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>